Port #642's loss changes to estimation.py

+++ estimation.py
@@ -103,7 +103,7 @@ def estimate_memory(job_config: JobConfig):
     # loss fn can be shared by pipeline-parallel or non-pp execution
     def loss_fn(pred, labels):
         return torch.nn.functional.cross_entropy(
-            pred.flatten(0, 1), labels.flatten(0, 1)
+            pred.flatten(0, 1).float(), labels.flatten(0, 1)
         )
 
     # build model (using meta init)
